Jenkins

Approved-by: Matthew Parno <mparno@mit.edu>

+++ SupportScripts/Jenkins/acdlBuildScript.sh
@@ -3,46 +3,41 @@
 shopt -s nocasematch
 
 #######################################
-##### DECIDE TO BUILD RELEASE OR DEBUG
+##### Use the directory name to determine build type
 #######################################
 dir=`pwd`
 bin_dir=$(echo "$dir" | tr '[:lower:]' '[:upper:]')
 
-build_type="Debug"
-if echo $bin_dir| grep -q "RELEASE"; then
-  build_type="Release"
+if echo $bin_dir| grep -q "PYTHON"; then
+  with_python=1
+  temp="${bin_dir##*PYTHON}"
+  PYTHON_VERSION="${temp:0:1}"
 else
-  build_type="Debug"
+  with_python=0
+  PYTHON_VERSION="NA"
 fi
 
 ####################################
 ##### SET MACHINE SPECIFIC PATHS
 ####################################
-
 if [[ `hostname` == "reynolds" ]]; then
 
   BOOST_SOURCE=/home/mparno/util/boost_1_63_0.tar.gz
-    
-  if echo $bin_dir| grep -q "CLANG"; then
-    GTEST_DIR=/home/mparno/util/gtest/clang
-    echo "Using gtest compiled with clang in ${GTEST_DIR}"
-  else
-    GTEST_DIR=/home/mparno/util/gtest/gnu
-    echo "Using gtest compiled with gcc in ${GTEST_DIR}"
-  fi
+  HDF5_SOURCE=/home/mparno/util/CMake-hdf5-1.8.19.tar.gz
+  NANOFLANN_SOURCE=/home/mparno/util/nanoflann-src.zip
+  NLOPT_SOURCE=/home/mparno/util/nlopt-2.4.2.tar.gz
+  STANMATH_SOURCE=/home/mparno/util/stanmath-v2.18.0.zip
+
+  GTEST_DIR=/home/mparno/util/gtest_install
 
 elif [[ `hostname` == "macys.mit.edu" ]]; then
 
   BOOST_SOURCE=/Users/mparno/util/boost_1_63_0.tar.gz
-    
+
   # export the path so we can get cmake
   export PATH=/usr/local/bin:/usr/local/sbin:$PATH
 
   GTEST_DIR=/Users/jenkins/util/gtest/
-else
-
-  # if not on macys or reynolds, assume I'm on Matt's macbook for testing
-  GTEST_DIR=/Users/mparno/Documents/Repositories/gtest_clang/
 
 fi
 
@@ -50,51 +45,8 @@ fi
 ##### EXTRACT COMPILER FROM WORKSPACE
 #######################################
 
-if echo $bin_dir| grep -q "CLANG38"; then
-
-  if [[ `hostname` == "reynolds" ]]; then
-    my_cc_compiler="clang-3.8"
-    my_cxx_compiler="clang++-3.8"
-  else
-    my_cc_compiler="clang"
-    my_cxx_compiler="clang++"
-  fi
-
-elif echo $bin_dir | grep -q "CLANG"; then
-
-  if [[ `hostname` == "macys.mit.edu" ]]; then
-    my_cc_compiler="clang"
-    my_cxx_compiler="clang++"
-  else
-    my_cc_compiler="clang-3.8"
-    my_cxx_compiler="clang++-3.8"
-  fi
-
-elif echo $bin_dir | grep -q "INTEL"; then
-  my_cc_compiler="icc"
-  my_cxx_compiler="icpc"
-elif echo $bin_dir | grep -q "GNU49"; then
-  my_cc_compiler="gcc-4.9"
-  my_cxx_compiler="g++-4.9"
-elif echo $bin_dir | grep -q "GNU48"; then
-  my_cc_compiler="gcc-4.8"
-  my_cxx_compiler="g++-4.8"
-elif echo $bin_dir | grep -q "GNU47"; then
-  my_cc_compiler="gcc-4.7"
-  my_cxx_compiler="g++-4.7"
-elif echo $bin_dir | grep -q "GNU5"; then
-  my_cc_compiler="gcc-5"
-  my_cxx_compiler="g++-5"
-elif echo $bin_dir | grep -q "GNU6"; then
-  my_cc_compiler="gcc-6"
-  my_cxx_compiler="g++-6"
-elif echo $bin_dir | grep -q "GNU7"; then
-  my_cc_compiler="gcc-7"
-  my_cxx_compiler="g++-7"
-elif echo $bin_dir | grep -q "GNU"; then
-  my_cc_compiler="gcc"
-  my_cxx_compiler="g++"
-fi
+my_cc_compiler="gcc"
+my_cxx_compiler="g++"
 
 echo "C Compiler = $my_cc_compiler"
 echo "CXX Compiler = $my_cxx_compiler"
@@ -117,23 +69,30 @@ fi
 
 # cd into build directory and remove all previous files
 cd "$BUILD_DIR"
-rm CMakeCache.txt
-rm -rf CMakeFiles
-rm -rf modules
+if [ -d "CMakeFiles" ]; then
+  rm CMakeCache.txt
+  rm -rf CMakeFiles
+  rm -rf modules
+fi
 
 #######################################
 ##### RUN CMAKE
 #######################################
 cmake \
--DCMAKE_BUILD_TYPE=$build_type \
+-DCMAKE_BUILD_TYPE=Release \
 -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR \
 -DCMAKE_CXX_COMPILER=$my_cxx_compiler \
 -DCMAKE_C_COMPILER=$my_cc_compiler \
 -DMUQ_USE_GTEST=ON \
 -DMUQ_GTEST_DIR=$GTEST_DIR \
--DMUQ_USE_OPENMPI=OFF \
 -DBOOST_EXTERNAL_SOURCE=$BOOST_SOURCE \
-$dir
+-DHDF5_EXTERNAL_SOURCE=$HDF5_SOURCE \
+-DNANOFLANN_EXTERNAL_SOURCE=$NANOFLANN_SOURCE \
+-DNLOPT_EXTERNAL_SOURCE=$NLOPT_SOURCE \
+-DSTANMATH_EXTERNAL_SOURCE=$STANMATH_SOURCE \
+-DMUQ_USE_PYTHON=$with_python \
+-DPYBIND11_PYTHON_VERSION=$PYTHON_VERSION \
+../
 
 #######################################
 ##### BUILD MUQ
@@ -141,6 +100,6 @@ $dir
 make install > OutputFromMake.txt
 tail -200 OutputFromMake.txt
 
-cd $dir
+cd "$dir"
 
 exit 0
